PPI/celery-docker/requirements.txt
gmpy2==2.0.8
celery==4.2.1
gevent==1.5
bitarray==1.2.2
pycryptodome==3.9.8
Flask==1.1.2
requests==2.24.0
simplejson
memory_profiler
psutil
numpy>=1.9.1
pandas
matplotlib
scipy